Rename metal_surface_energy module to surface_energy
Renamed the module to better reflect its support for both elemental metals
and stoichiometric intermetallics. Updated all import paths and references.

Changes:
- Rename teros/core/metal_surface_energy/ to teros/core/surface_energy/
- Update all import statements in example scripts and core __init__.py
- Update documentation in README.md and docs/metal_surface_energy.md
- Export identify_compound_type function from core module
